+++ object/src/main/java/com/_P/eureka/client/object/hub/dto/HubUpdateDto.java
@@ -8,7 +8,7 @@ public class HubUpdateDto {
 
     private String address;
 
-    private int latitude;
+    private Long latitude;
 
-    private int longitude;
+    private Long longitude;
 }
